LPS-101448 Moving DropDownItem to its own file. Main component renaming
