+++ modules/apps/asset/asset-list-web/src/main/resources/META-INF/resources/js/components/VariationsNav/index.js
